fix: renaming boost providers
