+++ src/swaps/liqualityboost/liqualityBoostERC20toNative/info.json
@@ -1,5 +1,5 @@
 {
-  "title": "LiqualityBoost Atomic Swaps (type 2)",
+  "title": "LiqualityBoost Atomic Swaps (ERC20 to Native)",
   "description": "The most secure, non-custodial way to swap across blockchains. Swap hundreds of erc20 tokens to any native asset ",
   "pros": [
     "Cross-chain",
