+++ src/swaps/liqualityboost/liqualityBoostNativeToERC20/info.json
@@ -1,5 +1,5 @@
 {
-  "title": "LiqualityBoost Atomic Swaps (type 1)",
+  "title": "LiqualityBoost Atomic Swaps (Native to ERC20)",
   "description": "The most secure, non-custodial way to swap across blockchains. Swap any native asset to hundreds of erc20 tokens ",
   "pros": [
     "Cross-chain",
